requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==4.6.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
